+++ lib/Random.js
@@ -9,6 +9,16 @@
  */
 class Random {
   /**
+   * Generate a random boolean.  Useful for binomial distributions.
+   *
+   * @param {number} pTrue
+   * @returns {boolean} `true` with probability `pTrue`, otherwise `false`
+   */
+  static bool(pTrue = 0.5) {
+    return Math.random() < pTrue;
+  }
+
+  /**
    * Generate a random number uniformly on the interval `[lo, hi)`.
    *
    * @param {number} lo - minimum value (inclusive)
