+++ lib/model/StudyRequestFilters.js
@@ -3,7 +3,7 @@ import Joi from '@/lib/model/Joi';
 
 export default {
   assignees: Joi.array().single().items(
-    Joi.enum().ofType(StudyRequestAssignee).allow('').required(),
+    Joi.enum().ofType(StudyRequestAssignee).allow(''),
   ).default([]),
   column: Joi.string().valid(
     'ASSIGNED_TO',
@@ -12,6 +12,7 @@ export default {
     'REQUESTER',
     'STATUS',
     'STUDY_TYPE',
+    null,
   ).default(null),
   limit: Joi
     .number()
@@ -20,7 +21,7 @@ export default {
     .max(100)
     .required(),
   offset: Joi.number().integer().min(0).required(),
-  query: Joi.string().default(null),
+  query: Joi.string().allow(null).default(null),
   sortBy: Joi.string().valid(
     'CREATED_AT',
     'DUE_DATE',
@@ -30,10 +31,10 @@ export default {
   ).required(),
   sortDesc: Joi.boolean().required(),
   statuses: Joi.array().single().items(
-    Joi.enum().ofType(StudyRequestStatus).required(),
+    Joi.enum().ofType(StudyRequestStatus),
   ).default([]),
   studyTypes: Joi.array().single().items(
-    Joi.enum().ofType(StudyType).required(),
+    Joi.enum().ofType(StudyType),
   ).default([]),
   studyTypeOther: Joi.boolean().default(false),
   userOnly: Joi.boolean().default(false),
