+++ lib/test/random/FilterGenerator.js
@@ -1,5 +1,11 @@
 import ArrayUtils from '@/lib/ArrayUtils';
-import { CollisionEmphasisArea, StudyHours, StudyType } from '@/lib/Constants';
+import {
+  CollisionEmphasisArea,
+  StudyHours,
+  StudyRequestAssignee,
+  StudyRequestStatus,
+  StudyType,
+} from '@/lib/Constants';
 import Random from '@/lib/Random';
 import DateTime from '@/lib/time/DateTime';
 
@@ -13,18 +19,38 @@ const factorValues = {
   rdsfcond: [1, 2, 3, 4, 99],
 };
 
+// UTILITY METHODS
+
+function generateSample(values) {
+  const applySample = Random.bool();
+  if (!applySample) {
+    return [];
+  }
+  const k = Random.choice(ks);
+  return Random.sample(values, k);
+}
+
+function generateEnumSample(enumClass, extraValues = []) {
+  const values = [
+    ...enumClass.enumValues,
+    ...extraValues,
+  ];
+  return generateSample(values);
+}
+
+// COLLISION FILTERS
+
 function generateMvcrFilter(fieldName) {
+  const applyMvcrFilter = Random.bool();
+  if (!applyMvcrFilter) {
+    return [];
+  }
   const k = Random.choice(ks);
   return Random.sample(factorValues[fieldName], k);
 }
 
-function generateEnumSample(enumClass) {
-  const k = Random.choice(ks);
-  return Random.sample(enumClass.enumValues, k);
-}
-
 function generateHoursOfDay() {
-  const filterHoursOfDay = Random.choice([true, false]);
+  const filterHoursOfDay = Random.bool();
   if (!filterHoursOfDay) {
     return [0, 24];
   }
@@ -61,6 +87,8 @@ function generateFiltersCollision() {
   };
 }
 
+// COMMON FILTERS
+
 function generateDateRange() {
   const now = DateTime.local();
   // TODO: implement this
@@ -78,7 +106,7 @@ function generateDateRange() {
 }
 
 function generateFiltersCommon() {
-  const applyDateRange = Random.choice([true, false]);
+  const applyDateRange = Random.bool();
   let dateRangeStart = null;
   let dateRangeEnd = null;
   if (applyDateRange) {
@@ -98,6 +126,8 @@ function generateFiltersCommon() {
   };
 }
 
+// STUDY FILTERS
+
 function generateFiltersStudy() {
   const hours = generateEnumSample(StudyHours);
   const studyTypes = generateEnumSample(StudyType);
@@ -115,6 +145,54 @@ function generateFilters() {
   };
 }
 
+function generateQueryFilter() {
+  const applyQuery = Random.bool();
+  if (!applyQuery) {
+    return { column: null, query: null };
+  }
+  const column = Random.choice([
+    'ASSIGNED_TO',
+    'ID',
+    'LOCATION',
+    'REQUESTER',
+    'STATUS',
+    'STUDY_TYPE',
+    null,
+  ]);
+  const query = column === 'ID' ? 73 : 'ATR';
+  return { column, query };
+}
+
+function generateFiltersStudyRequest() {
+  const assignees = generateEnumSample(StudyRequestAssignee, ['']);
+  const queryFilter = generateQueryFilter();
+  const sortBy = Random.choice([
+    'CREATED_AT',
+    'DUE_DATE',
+    'ID',
+    'LOCATION',
+    'REQUESTER',
+  ]);
+  const sortDesc = Random.bool();
+  const statuses = generateEnumSample(StudyRequestStatus);
+  const studyTypes = generateEnumSample(StudyType);
+  const studyTypeOther = Random.bool();
+  const userOnly = Random.bool();
+
+  return {
+    assignees,
+    limit: 10,
+    offset: 0,
+    sortBy,
+    sortDesc,
+    statuses,
+    studyTypes,
+    studyTypeOther,
+    userOnly,
+    ...queryFilter,
+  };
+}
+
 /**
  * @namespace
  */
@@ -123,6 +201,7 @@ const FilterGenerator = {
   generateFiltersCollision,
   generateFiltersCommon,
   generateFiltersStudy,
+  generateFiltersStudyRequest,
 };
 
 export {
@@ -131,4 +210,5 @@ export {
   generateFiltersCollision,
   generateFiltersCommon,
   generateFiltersStudy,
+  generateFiltersStudyRequest,
 };
